html_parser/source_code/requirements.txt
requests==2.26.0
google-cloud-storage==1.43.0
bs4==0.0.1
tqdm==4.62.3
